Rename eval.py to predict.py. Use of new 'experiments/' file structure.
Predict.py is now simpler to use as it reads the meta file created by
train.py to load in model architecture args
